src/requirements.txt
opencv-python==4.5.3.56
Pillow==8.4.0
tk==0.1.0